requirements.txt
grpcio
cffi
Flask
Flask-Security-Too
Flask-Caching
python-dotenv
Flask-SQLAlchemy
Flask-Babel
email-validator
bcrypt
cryptography
PyQRCode
phonenumberslite
zxcvbn
bleach
google-analytics-data
mailjet-rest